+++ docs/CHANGELOG.rst
@@ -19,6 +19,7 @@ Added
 - Add relations property to ``data:`` and ``list:data:`` fields to support
   relations on client
 - Add ``entity_id`` property do ``DataField`` in Python processes
+- Add relations in Python processes
 
 
 ===================